pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "thesis"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "blosc>=1.11.3",
    # cuml-cu12==25.8.0  # Linux-only - managed separately
    "cupy-cuda12x>=13.6.0",
    "dask>=0.17.5",
    "fiona>=1.10.1",
    "geopandas>=1.1.1",
    "matplotlib>=3.10.3",
    "memory-profiler>=0.61.0",
    "numpy==2.2.*",
    "optuna>=4.5.0",
    "pandas>=2.3.1",
    "plotly>=6.3.0",
    "pynvml>=12.0.0",
    "rasterio>=1.4.3",
    "scikit-image>=0.24.0",
    "scikit-learn>=1.7.1",
    "statsmodels>=0.14.5",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "xgboost>=2.1.3",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"

[[tool.uv.index]]
url = "https://pypi.nvidia.com/"

[tool.uv]
environments = ["sys_platform == 'linux'"]
index-strategy = "unsafe-best-match"

# Note: cuml-cu12 is Linux-only and excluded from dependencies
# Install on Linux server with: pip install --extra-index-url https://pypi.nvidia.com/ cuml-cu12==25.8.0

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
]
ignore = [
    "E402",  # Module level import not at top of file 
    "E731",  # def <-> lambda assignments, used this way too much in some contexts
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__ files
"test_*.py" = ["F401"]    # Allow unused imports in test files
"researchCode/*.py" = ["E402", "E731"]  # Research code can be more flexible

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
